ingestion/requirements.txt
chromadb==0.4.24
numpy<2
openai>=1.3.0
pandas==2.2.2
pyarrow==16.1.0
pydantic==2.7.0
pyyaml==6.0.1
tiktoken==0.6.0
typer==0.12.3
